scripts/requirements.txt
torch
torchvision==0.20.0
lmdb
fastapi[standard]
streamlit
pillow